+++ packages/uploader/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.60 (January 05, 2024)
+
+- chore: polish resources  middleware
+
 ## 0.0.59 (January 03, 2024)
 
 - chore(resource): update resource export folder and Blocklet SDK